# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Resolvy API Documentation",
        "description": "API documentation for the Resolvy helpdesk application",
        "contact": {
            "email": "admin@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Main API Server"
        }
    ],
    "paths": {
...
        "/api/v1/tickets": {
            "get": {
                "tags": [
                    "Tickets"
                ],
                "summary": "Get list of tickets",
                "description": "Returns a paginated list of tickets",
                "operationId": "getTicketsList",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "The page number",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/TicketResource"
                                            }
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Tickets retrieved successfully."
                                        },
                                        "meta": {
                                            "$ref": "#/components/schemas/PaginationMeta"
                                        },
                                        "links": {
                                            "$ref": "#/components/schemas/PaginationLinks"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Tickets"
                ],
                "summary": "Create a new ticket",
                "description": "Creates a new ticket record",
                "operationId": "storeTicket",
                "requestBody": {
                    "description": "Ticket data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreTicketRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Ticket created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/TicketResource"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Ticket created successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/tickets/{id}": {
            "get": {
                "tags": [
                    "Tickets"
                ],
                "summary": "Get ticket information",
                "description": "Returns ticket data",
                "operationId": "getTicketById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of ticket to return",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/TicketResource"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Ticket retrieved successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Resource not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Tickets"
                ],
                "summary": "Update an existing ticket",
                "description": "Updates a ticket record",
                "operationId": "updateTicket",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of ticket to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Ticket data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateTicketRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Ticket updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/TicketResource"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Ticket updated successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Tickets"
                ],
                "summary": "Delete an existing ticket",
                "description": "Deletes a ticket record",
                "operationId": "deleteTicket",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of ticket to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ticket deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "example": null
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Ticket deleted successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/tickets/{id}/status": {
            "post": {
                "tags": [
                    "Tickets"
                ],
                "summary": "Update a ticket's status",
                "operationId": "updateTicketStatus",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateTicketStatusRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Status updated successfully"
                    },
                    "422": {
                        "description": "Status update failed due to validation or workflow rule violation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/tickets/{id}/assignments": {
            "post": {
                "tags": [
                    "Tickets"
                ],
                "summary": "Update a ticket's assigned users",
                "operationId": "updateTicketAssignments",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateTicketAssignmentsRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Assignments updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/UserResource"
                                            }
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Assignments updated successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "PaginationMeta": {
                "title": "Pagination Meta",
                "properties": {
                    "current_page": {
                        "type": "integer",
                        "example": 1
                    },
                    "from": {
                        "type": "integer",
                        "example": 1
                    },
                    "last_page": {
                        "type": "integer",
                        "example": 10
                    },
                    "path": {
                        "type": "string",
                        "example": "http://localhost:8000/api/v1/customers"
                    },
                    "per_page": {
                        "type": "integer",
                        "example": 15
                    },
                    "to": {
                        "type": "integer",
                        "example": 15
                    },
                    "total": {
                        "type": "integer",
                        "example": 150
                    }
                },
                "type": "object"
            },
            "PaginationLinks": {
                "title": "Pagination Links",
                "properties": {
                    "first": {
                        "type": "string",
                        "example": "http://localhost:8000/api/v1/customers?page=1"
                    },
                    "last": {
                        "type": "string",
                        "example": "http://localhost:8000/api/v1/customers?page=10"
                    },
                    "prev": {
                        "type": "string",
                        "example": null,
                        "nullable": true
                    },
                    "next": {
                        "type": "string",
                        "example": "http://localhost:8000/api/v1/customers?page=2"
                    }
                },
                "type": "object"
            },
...
            "UserResource": {
                "title": "User Resource",
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": true,
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "john.doe@example.com"
                    }
                },
                "type": "object"
            },
            "StoreTicketRequest": {
                "title": "Store Ticket Request",
                "required": [
                    "subject",
                    "customer_id",
                    "priority"
                ],
                "properties": {
                    "subject": {
                        "description": "The subject of the ticket",
                        "type": "string",
                        "example": "Cannot connect to the server"
                    },
                    "description": {
                        "description": "Detailed description of the issue",
                        "type": "string",
                        "example": "When I try to log in, I get a connection error message."
                    },
                    "customer_id": {
                        "description": "ID of the customer reporting the issue",
                        "type": "integer",
                        "example": 1
                    },
                    "ticket_category_id": {
                        "description": "ID of the ticket category",
                        "type": "integer",
                        "example": 1
                    },
                    "priority": {
                        "type": "string",
                        "enum": [
                            "low",
                            "medium",
                            "high"
                        ],
                        "example": "medium"
                    },
                    "assigned_tos": {
                        "description": "An array of user IDs to assign to the ticket",
                        "type": "array",
                        "items": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                },
                "type": "object"
            },
            "UpdateTicketRequest": {
                "title": "Update Ticket Request",
                "properties": {
                    "subject": {
                        "description": "The subject of the ticket",
                        "type": "string",
                        "example": "Cannot connect to the server"
                    },
                    "description": {
                        "description": "Detailed description of the issue",
                        "type": "string",
                        "example": "When I try to log in, I get a connection error message."
                    },
                    "customer_id": {
                        "description": "ID of the customer reporting the issue",
                        "type": "integer",
                        "example": 1
                    },
                    "ticket_category_id": {
                        "description": "ID of the ticket category",
                        "type": "integer",
                        "example": 1
                    },
                    "priority": {
                        "type": "string",
                        "enum": [
                            "low",
                            "medium",
                            "high"
                        ],
                        "example": "medium"
                    },
                    "assigned_tos": {
                        "description": "An array of user IDs to assign to the ticket",
                        "type": "array",
                        "items": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                },
                "type": "object"
            },
            "UpdateTicketStatusRequest": {
                "title": "Update Ticket Status Request",
                "required": [
                    "status"
                ],
                "properties": {
                    "status": {
                        "type": "string",
                        "enum": [
                            "open",
                            "in_progress",
                            "resolved",
                            "closed"
                        ],
                        "example": "in_progress"
                    }
                },
                "type": "object"
            },
            "UpdateTicketAssignmentsRequest": {
                "title": "Update Ticket Assignments Request",
                "required": [
                    "assigned_tos"
                ],
                "properties": {
                    "assigned_tos": {
                        "description": "An array of user IDs to assign. Send an empty array to unassign all.",
                        "type": "array",
                        "items": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                },
                "type": "object"
            },
            "TicketResource": {
                "title": "Ticket Resource",
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": true,
                        "example": 1
                    },
                    "ticket_no": {
                        "type": "string",
                        "readOnly": true,
                        "example": "SUP-0000000001"
                    },
                    "subject": {
                        "type": "string",
                        "example": "Cannot connect to the server"
                    },
                    "description": {
                        "type": "string",
                        "example": "When I try to log in..."
                    },
                    "status": {
                        "type": "string",
                        "example": "open"
                    },
                    "priority": {
                        "type": "string",
                        "example": "medium"
                    },
                    "customer": {
                        "$ref": "#/components/schemas/CustomerResource"
                    },
                    "subscription": {
                        "$ref": "#/components/schemas/SubscriptionResource"
                    },
                    "ticket_category": {
                        "$ref": "#/components/schemas/TicketCategoryResource"
                    },
                    "assigned_tos": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/UserResource"
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}